Class ADEg.C1

An object of ADEg.C1 class represents unidimensional data into two dimensions.

The ADEg.C1 class is a virtual class, i.e. a class which is not possible to create objects but which have heirs. This class inherits from ADEg class and has three son classes : C1.barchart, C1.curve, C1.density, C1.dotplot, C1.gauss, C1.hist, C1.interval

Objects from the Class

None object of this class can be instantiated.

Slots

data

a list containing data or data's name.

score: the displayed values in the form of a numeric vector, a name or a matching call.

frame: a positive or null integer. It is the number of the frame containing the data (used with sys.frame(..., env = data$frame)). Only if the data are not stored (i.e. data$storeData = FALSE).

storeData: a logical indicating if the data should be stored in the returned object. If FALSE, only the names of the data arguments are stored.

trellis.par

a list of parameters for lattice call. It will be passed directly to par.settings arguments of the lattice function.

adeg.par

a list of graphical parameters, corresponding to the ones given by adegpar() function.

lattice.call

a list of two elements to create the trellis object:

graphictype: xyplot

arguments: its parameters to obtain the trellis object

g.args

a list containing some method parameters linked with the created object of ADEg.C1 class

stats

a list of internal preliminary calculations

s.misc

a list of some others internal parameters:

hori.update: a logical indicating if the sense of direction of the graphics is updating

backgrid: a list of two elements for grid lines. backgrid$x defines the coordinates of the lines (horizontal or vertical depending on the graphics orientation) and backgrid$d the grid mesh

Call

an object of class call

Extends

Class ADEg, directly.

Methods

prepare

signature(object = "ADEg.C1"): performs the calculations before display the object (e.g. limits, grid and axis calculations)

setlatticecall

signature(object = "ADEg.C1"): prepares the lattice.call slot

panelbase

signature(object = "ADEg.C1"): defines the graphical background (e.g. grid, rugs and box)

gettrellis

signature(object = "ADEg.C1"): converts the graphic into a trellis object of lattice class

Note

The ADEg.S1 class and ADEg.C1 class are both used to represent an unidimensional information (e.g. a score). The difference between these two classes is mainly ideological : an ADEg.S1 object is a representation into one dimension (e.g. one line) while an ADEg.C1 object is a representation into two dimensions (e.g. curves).

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

See also

Examples

showClass("ADEg.C1")
#> Virtual Class "ADEg.C1" [package "adegraphics"]
#> 
#> Slots:
#>                                                                        
#> Name:          data  trellis.par     adeg.par lattice.call       g.args
#> Class:         list         list         list         list         list
#>                                              
#> Name:         stats       s.misc         Call
#> Class:         list         list         call
#> 
#> Extends: 
#> Class "ADEg", directly
#> Class "ADEgORtrellis", by class "ADEg", distance 2
#> Class "ADEgORADEgSORtrellis", by class "ADEg", distance 2
#> 
#> Known Subclasses: 
#> Class "C1.barchart", directly
#> Class "C1.curve", directly
#> Class "C1.density", directly
#> Class "C1.gauss", directly
#> Class "C1.dotplot", directly
#> Class "C1.hist", directly
#> Class "C1.interval", directly
#> Class "C1.curves", by class "C1.curve", distance 2