Class ADEg.S2

An object of ADEg.S2 class represents bi-dimensional data.

The ADEg.S2 class is a virtual class, i.e. a class which is not possible to create objects but which have heirs. This class inherits from ADEg class and has eleven son classes : S2.arrow, S2.class, S2.corcircle, S2.density, S2.distri, S2.image, S2.label, S2.logo, S2.match, S2.traject and S2.value.

Objects from the Class

None object of this class can be instantiated.

Slots

data

a list containing data or data's name.

dfxy: the displayed values in the form of a data frame, a name or a matching call.

xax: an integer or a vector indicating the columns of dfxy kept for the x-axes.

yax: an integer or a vector indicating the columns of dfxy kept for the y-axes.

frame: a positive or null integer. It is the number of the frame containing the data (used with sys.frame(..., env = data$frame)). Only if the data are not stored (i.e. data$storeData = FALSE).

storeData: a logical indicating if the data should be stored in the returned object. If FALSE, only the names of the data arguments are stored.

trellis.par

a list of parameters for lattice call. It will be passed directly to par.settings arguments of the lattice function.

adeg.par

a list of graphical parameters, corresponding to the ones given by adegpar() function.

lattice.call

a list of two elements to create the trellis object:

graphictype: xyplot

arguments: its parameters to obtain the trellis object

g.args

a list containing some method parameters linked with the created object of ADEg.S2 class:

fullcircle: only for S2.corcircle objects

method: only for S2.value objects

symbol: only for S2.value objects

center: only for S2.value objects

stats

a list of internal preliminary calculations

s.misc

a list of some others internal parameters:

xfullcircle.update and yfullcircle.update: a logical indicating if the circle size is updating (only for S2.corcircle objects)

plegend.update: a logical indicating if the legend parameters are updating

breaks.update: a logical indicating if the legend breaks are updating

backgrid: a list of elements for grid lines

Call

an object of class call

Extends

Class ADEg, directly.

Methods

prepare

signature(object = "ADEg.S2"): performs the calculations before display the object (e.g. limits, grid and axis calculations)

setlatticecall

signature(object = "ADEg.S2"): prepares the lattice.call slot

panelbase

signature(object = "ADEg.S2"): defines the graphical background (e.g. grid and box)

gettrellis

signature(object = "ADEg.S2"): converts the graphic into a trellis object of lattice class

zoom

signature(object = "ADEg.S2", zoom = "numeric", center = "missing"): performs a zoom in (if zoom < 1) or out (if zoom > 1) centered

zoom

signature(object = "ADEg.S2", zoom = "numeric", center = "numeric"): performs a zoom in (if zoom < 1) or out (if zoom > 1) around the center passed in parameter (center should be a two-length vector)

addhist

signature(object = "ADEg.S2"): adds histograms and density lines against a bi-dimensional graphics

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

See also

Examples

showClass("ADEg.S2")
#> Virtual Class "ADEg.S2" [package "adegraphics"]
#> 
#> Slots:
#>                                                                        
#> Name:          data  trellis.par     adeg.par lattice.call       g.args
#> Class:         list         list         list         list         list
#>                                              
#> Name:         stats       s.misc         Call
#> Class:         list         list         call
#> 
#> Extends: 
#> Class "ADEg", directly
#> Class "ADEgORtrellis", by class "ADEg", distance 2
#> Class "ADEgORADEgSORtrellis", by class "ADEg", distance 2
#> 
#> Known Subclasses: "S2.arrow", "S2.class", "S2.corcircle", "S2.density", "S2.distri", "S2.image", 
#> "S2.label", "S2.logo", "S2.match", "S2.traject", "S2.value"