Class C1.curve

A class for the creation and display of a numeric score linked by curves. The C1.curves allows to deal with multiple scores.

Objects from the Class

C1.curve objects can be created by calls of the form new("C1.curve", ...). The regular usage in this package is to use the s1d.curve function.

C1.curves objects can be created by calls of the form new("C1.curves", ...). The regular usage in this package is to use the s1d.curves function. Class C1.curves extends C1.curve directly.

Slots

data

a list containing data or data's name.

score: the displayed values in the form of a vector, a factor, a name or a matching call.

at: the index value.

frame: a positive or null integer. It is the number of the frame containing the data (used with sys.frame(..., env = data$frame)). Only if the data are not stored (i.e. data$storeData = FALSE).

storeData: a logical indicating if the data should be stored in the returned object. If FALSE, only the names of the data arguments are stored.

trellis.par

a list of parameters for lattice call. It will be passed directly to par.settings arguments of the lattice function.

adeg.par

a list of graphical parameters, corresponding to the ones given by adegpar() function.

lattice.call

a list to create the trellis object.

g.args

a list containing some method parameters linked with the created object of ADEg.C1 class.

stats

a list of internal preliminary calculations

s.misc

a list of some others internal parameters

Call

an object of class call

Extends

Class ADEg.C1, directly.
Class ADEg, by class ADEg.C1, distance 2.
Class ADEgORtrellis, by class ADEg.C1, distance 3.
Class ADEgORADEgSORtrellis, by class ADEg.C1, distance 3.

Methods

The methods of the father classes "ADEg.C1" and "ADEg" can be used by inheritance. The specific methods for C1.curve and C1.curves are:

prepare

signature(object = "C1.curve"): calls the parent method (prepare for ADEg.C1) and modifies some graphical parameters used by default.

panel

signature(object = "C1.curve"): draws points and curves.

panel

signature(object = "C1.curves"): draws points and curves.

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

See also

Examples

showClass("C1.curve")
#> Class "C1.curve" [package "adegraphics"]
#> 
#> Slots:
#>                                                                        
#> Name:          data  trellis.par     adeg.par lattice.call       g.args
#> Class:         list         list         list         list         list
#>                                              
#> Name:         stats       s.misc         Call
#> Class:         list         list         call
#> 
#> Extends: 
#> Class "ADEg.C1", directly
#> Class "ADEg", by class "ADEg.C1", distance 2
#> Class "ADEgORtrellis", by class "ADEg.C1", distance 3
#> Class "ADEgORADEgSORtrellis", by class "ADEg.C1", distance 3
#> 
#> Known Subclasses: "C1.curves"
showClass("C1.curves")
#> Class "C1.curves" [package "adegraphics"]
#> 
#> Slots:
#>                                                                        
#> Name:          data  trellis.par     adeg.par lattice.call       g.args
#> Class:         list         list         list         list         list
#>                                              
#> Name:         stats       s.misc         Call
#> Class:         list         list         call
#> 
#> Extends: 
#> Class "C1.curve", directly
#> Class "ADEg.C1", by class "C1.curve", distance 2
#> Class "ADEg", by class "C1.curve", distance 3
#> Class "ADEgORtrellis", by class "C1.curve", distance 4
#> Class "ADEgORADEgSORtrellis", by class "C1.curve", distance 4