Class C1.interval

A class for the creation and display of an interval between two numeric scores.

Objects from the Class

C1.interval objects can be created by calls of the form new("C1.interval", ...).

The regular usage in this package is to use the s1d.interval function.

Slots

data

a list containing data or data's name.

score: the displayed values in the form of a vector, a factor, a name or a matching call.

at: the index value.

frame: a positive or null integer. It is the number of the frame containing the data (used with sys.frame(..., env = data$frame)). Only if the data are not stored (i.e. data$storeData = FALSE).

storeData: a logical indicating if the data should be stored in the returned object. If FALSE, only the names of the data arguments are stored.

trellis.par

a list of parameters for lattice call. It will be passed directly to par.settings arguments of the lattice function.

adeg.par

a list of graphical parameters, corresponding to the ones given by adegpar() function.

lattice.call

a list to create the trellis object.

g.args

a list containing some method parameters linked with the created object of ADEg.C1 class. The specific slot for C1.density objects is:

method: a value, bars or area, to represent either segments or areas between scores.

stats

a list of internal preliminary calculations

s.misc

a list of some others internal parameters

Call

an object of class call

Extends

Class ADEg.C1, directly.
Class ADEg, by class ADEg.C1, distance 2.
Class ADEgORtrellis, by class ADEg.C1, distance 3.
Class ADEgORADEgSORtrellis, by class ADEg.C1, distance 3.

Methods

The methods of the father classes "ADEg.C1" and "ADEg" can be used by inheritance. The specific methods for C1.interval are:

prepare

signature(object = "C1.interval"): calls the parent method (prepare for ADEg.C1) and modifies some graphical parameters used by default.

panel

signature(object = "C1.interval"): draws segments or polygons.

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

See also

Examples

showClass("C1.interval")
#> Class "C1.interval" [package "adegraphics"]
#> 
#> Slots:
#>                                                                        
#> Name:          data  trellis.par     adeg.par lattice.call       g.args
#> Class:         list         list         list         list         list
#>                                              
#> Name:         stats       s.misc         Call
#> Class:         list         list         call
#> 
#> Extends: 
#> Class "ADEg.C1", directly
#> Class "ADEg", by class "ADEg.C1", distance 2
#> Class "ADEgORtrellis", by class "ADEg.C1", distance 3
#> Class "ADEgORADEgSORtrellis", by class "ADEg.C1", distance 3