Class S2.corcircle

A class for creating and drawing a correlation circle.

Objects from the Class

S2.corcircle objects can be created by calls of the form new("S2.corcircle", ...).

The regular usage in this package is to use the s.corcircle function.

Slots

data

a list containing data or data's name.

dfxy: the displayed values in the form of a data frame, a name or a matching call.

xax: an integer or a vector indicating the columns of dfxy kept for the x-axes.

yax: an integer or a vector indicating the columns of dfxy kept for the y-axes.

labels: a vector containing the points' labels.

frame: a positive or null integer. It is the number of the frame containing the data (used with sys.frame(..., env = data$frame)). Only if the data are not stored (i.e. data$storeData = FALSE).

storeData: a logical indicating if the data should be stored in the returned object. If FALSE, only the names of the data arguments are stored.

trellis.par

a list of parameters for lattice call. It will be passed directly to par.settings arguments of the lattice function.

adeg.par

a list of graphical parameters, corresponding to the ones given by adegpar() function.

lattice.call

a list to create the trellis object.

g.args

a list containing some method parameters linked with the created object of ADEg.S2 class. The specific slot for S2.corcircle objects is:

fullcircle: a logical to include the complete circle (limits are then c(-1, 1)).

stats

a list of internal preliminary calculations

s.misc

a list of some others internal parameters:

backgrid: a list of elements for grid lines

Call

an object of class call

Extends

Class ADEg.S2, directly.
Class ADEg, by class ADEg.S2, distance 2.
Class ADEgORtrellis, by class ADEg.S2, distance 3.
Class ADEgORADEgSORtrellis, by class ADEg.S2, distance 3.

Methods

The methods of the father classes "ADEg.S2" and "ADEg" can be used by inheritance. The specific methods for S2.corcircle are:

prepare

signature(object = "S2.corcircle"): calls the parent method (prepare for ADEg.S2), modifies some graphical parameters used by default and prepares the drawn grid.

panel

signature(object = "S2.corcircle"): draws arrows, labels and axes.

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

See also

Examples

showClass("S2.corcircle")
#> Class "S2.corcircle" [package "adegraphics"]
#> 
#> Slots:
#>                                                                        
#> Name:          data  trellis.par     adeg.par lattice.call       g.args
#> Class:         list         list         list         list         list
#>                                              
#> Name:         stats       s.misc         Call
#> Class:         list         list         call
#> 
#> Extends: 
#> Class "ADEg.S2", directly
#> Class "ADEg", by class "ADEg.S2", distance 2
#> Class "ADEgORtrellis", by class "ADEg.S2", distance 3
#> Class "ADEgORADEgSORtrellis", by class "ADEg.S2", distance 3