Methods prepare for ADEg objects

The method prepare performs the first calculus needed for the display.

Methods

signature(object = "ADEg.C1")

performs the calculations before display the object (e.g. limits, grid and axis calculations)

signature(object = "C1.barchart")

calls the parent method (prepare for ADEg.C1) and modifies some graphical parameters used by default

signature(object = "C1.curve")

calls the parent method (prepare for ADEg.C1) and modifies some graphical parameters used by default

signature(object = "C1.density")

calls the parent method (prepare for ADEg.C1), modifies some graphical parameters used by default and calculates the density curves according to the numeric score and the values' categories

signature(object = "C1.dotplot")

calls the parent method (prepare for ADEg.C1) and modifies some graphical parameters used by default

signature(object = "C1.gauss")

calls the parent method (prepare for ADEg.C1), modifies some graphical parameters used by default and calculates the Gauss curves according to the numeric score and the values' categories (using weighted mean and standard deviation)

signature(object = "C1.hist")

calls the parent method (prepare for ADEg.C1), modifies some graphical parameters used by default and calculates the boundaries and the height of cells

signature(object = "C1.interval")

calls the parent method (prepare for ADEg.C1) and modifies some graphical parameters used by default

signature(object = "ADEg.S1")

performs the calculations before display the object (e.g. limits, grid and axis calculations)

signature(object = "S1.boxplot")

calls the parent method (prepare for ADEg.S1) and modifies some graphical parameters used by default

signature(object = "S1.class")

calls the parent method (prepare for ADEg.S1) and modifies some graphical parameters used by default

signature(object = "S1.distri")

calls the parent method (prepare for ADEg.S1), modifies some graphical parameters used by default and calculates weighted mean and standard deviation

signature(object = "S1.label")

calls the parent method (prepare for ADEg.S1) and modifies some graphical parameters used by default

signature(object = "S1.match")

calls the parent method (prepare for ADEg.S1) and modifies some graphical parameters used by default

signature(object = "ADEg.S2")

performs the calculations before display the object (e.g. limits, grid and axis calculations)

signature(object = "S2.arrow")

calls the parent method (prepare for ADEg.S2), modifies some graphical parameters used by default and calculates limits

signature(object = "S2.class")

calls the parent method (prepare for ADEg.S2), modifies some graphical parameters used by default and calculates ellipses, convex hulls and centroids

signature(object = "S2.corcircle")

calls the parent method (prepare for ADEg.S2), modifies some graphical parameters used by default and prepares the drawn grid

signature(object = "S2.density")

calls the parent method (prepare for ADEg.S2), modifies some graphical parameters used by default and calculates densities

signature(object = "S2.distri")

calls the parent method (prepare for ADEg.S2), modifies some graphical parameters used by default and calculates ellipses and centroids

signature(object = "S2.image")

calls the parent method (prepare for ADEg.S2), modifies some graphical parameters used by default and calculates grid expansion and limits

signature(object = "S2.label")

calls the parent method (prepare for ADEg.S2) and modifies some graphical parameters used by default

signature(object = "S2.logo")

calls the parent method (prepare for ADEg.S2) and modifies some graphical parameters used by default

signature(object = "S2.match")

calls the parent method (prepare for ADEg.S2) and modifies some graphical parameters used by default

signature(object = "S2.traject")

calls the parent method (prepare for ADEg.S2) and modifies some graphical parameters used by default

signature(object = "S2.value")

calls the parent method (prepare for ADEg.S2), modifies some graphical parameters used by default and calculates limits

signature(object = "ADEg.T")

performs the calculations before display the object (e.g. limits, grid and axis calculations)

signature(object = "T.image")

calls the parent method (prepare for ADEg.T) and modifies some graphical parameters used by default and calculates limits and grid

signature(object = "T.value")

calls the parent method (prepare for ADEg.T) and modifies some graphical parameters used by default and calculates limits and grid

signature(object = "ADEg.Tr")

performs the calculations before display the object (e.g. limits, grid and axis calculations)

signature(object = "Tr.class")

calls the parent method (prepare for ADEg.Tr), modifies some graphical parameters used by default and calculated ellipses, convex hulls and centroids

signature(object = "Tr.label")

calls the parent method (prepare for ADEg.Tr) and modifies some graphical parameters used by default

signature(object = "Tr.match")

calls the parent method (prepare for ADEg.Tr), modifies some graphical parameters used by default and defines the mean point and the axis

signature(object = "Tr.traject")

calls the parent method (prepare for ADEg.Tr) and modifies some graphical parameters used by default

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray