Sort a sequence of graphical parameters

Sort a sequence of graphical parameters in several lists.

Usage

sortparamADEg(...)
sortparamADEgS(..., graphsnames, nbsubgraphs = rep(1, length(graphsnames)))

Arguments

...

a sequence of graphical parameters

graphsnames

a sequence containing the name of each simple graph of the ADEgS

nbsubgraphs

a sequence containing the number of sub-graphs in each graph named in graphsnames

Value

sortparamADEg return a list of four lists named adepar, trellis, g.args and rest. sortparamADEgS return a list of as many lists as the length of graphsnames, i.e., as the number of sub-graphs of the ADEgS. The names of the lists are graphsnames and each sub-list is the result of the sortparamADEg function aplly on each sub-graph.

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

Examples

l1 <- sortparamADEg(xlab = "x-axis label", ylab = "y-axis label", plabels.cex = 1.5, 
  porigin.include = FALSE)
length(l1)
#> [1] 4
names(l1)
#> [1] "adepar"  "trellis" "g.args"  "rest"   
  
l2 <- sortparamADEgS(xlab = "x-axis label", eig.main = "Eigenvalues", row.ppoints.col = "red",
  porigin.include = FALSE, graphsnames = c("row", "col", "eig"))
names(l2)
#> [1] "row" "col" "eig"
names(l2$row)
#> [1] "ppoints" "porigin" "xlab"   

l3 <- sortparamADEgS(xlab = "x-axis label", eig.main = "Eigenvalues", row.ppoints.col = "pink",
  porigin.include = FALSE, graphsnames = c("row", "col", "eig"), nbsubgraphs = c(1, 2, 1))
names(l3)
#> [1] "row" "col" "eig"
length(l3$row)
#> [1] 3
length(l3$col)
#> [1] 2