Class ADEg

An object of ADEg class is a simple graphic. This object can be blended in with another one (superposition, insertion and/or juxtaposition) to form a more complex graphics (an ADEgS object).

The ADEg class is a virtual class, i.e. a class which is not possible to create objects but which have heirs. This class has five son classes : ADEg.S1, ADEg.S2, ADEg.C1, ADEg.T and ADEg.Tr.

Objects from the Class

None object of this class can be instantiated.

Slots

trellis.par

a list of parameters for lattice call. It will be passed directly to par.settings arguments of the lattice function.

adeg.par

a list of graphical parameters, corresponding to the ones given by adegpar() function.

lattice.call

a list of two elements to create the trellis object:

graphictype: the lattice function to use

arguments: its parameters to obtain the trellis object

g.args

a list containing some parameters linked with the created object of ADEg class:

xlim, ylim

main, sub

xlab, ylab

samelimits

scales: a list of scales informations (ticks, marks and labels for the x-axis or the y-axis) in the form of the lattice argument scales in the xyplot function

stats

a list of internal preliminary calculations

s.misc

a list of some other internal parameters

Call

an object of class call

Methods

panelbase

signature(object = "ADEg"): draws grid and text and produces graphical output from the graphical object

getcall

signature(object = "ADEg"): returns the Call slot

getlatticecall

signature(object = "ADEg"): returns the lattice.call slot

getstats

signature(object = "ADEg"): returns the stats slot

getparameters

signature(object = "ADEg", number): if number is 1, returns the trellis.par slot, if it is 2, returns the adeg.par slot and if it is 0, returns the both slots

add.ADEg

signature(object = "ADEg"): superposes an ADEg on the current one plotted

+

signature(e1 = "ADEg", e2 = "ADEg"): superposes e2 on e1

superpose

signature(g1 = "ADEgORtrellis", g2 = "ADEgORtrellis", which = "ANY", plot = "ANY"): creates a new ADEgS object performing a superposition of g2 on g1.

printSuperpose

signature(g1 = "ADEgORtrellis", refg = "ADEgORtrellis"): internal method, not for users.

cbindADEg

signature(g1 = "ADEgORADEgS", g2 = "ADEgORADEgS"): creates a new "ADEgS" object combining g1 on g2.

rbindADEg

signature(g1 = "ADEgORADEgS", g2 = "ADEgORADEgS"): creates a new "ADEgS" object combining g1 on g2 by rows.

insert

signature(graphics = "ADEgORtrellis", oldgraphics = "missing", posi, ratio, inset, plot, which): creates a new ADEgS object performing an insertion of graphics into the current device.

insert

signature(graphics = "ADEgORtrellis", oldgraphics = "ADEg", posi, ratio, inset, plot): creates a new ADEgS object performing an insertion of graphics into oldgraphics.

show

signature(x = "ADEg"): prints the ADEg object

plot

signature(x = "ADEg"): prints the ADEg object

print

signature(x = "ADEg"): displays the ADEg object in the current device or in a new one

update

signature(object = "ADEg"): modifies graphical parameters after the ADEg creation, updates the current display and returns the modified ADEg

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

Note

For any ADEg creation, various graphical parameters can be passed into the dots (...) arguments.

the parameters listed in adegpar() can be changed, even if some of them do no modify the graphic representation chosen.

the lattice parameters listed in trellis.par.get() can also be changed.

limits, main and sub title, and axes labels can be changed using the keys xlim, ylim, main, sub, xlab and ylab.

a neighbouring graph (object of class nb or listw) and a spatial one (object of class sp) can be display in the background using the keys nbobject, Sp and sp.layout.

See also

Examples

showClass("ADEg")
#> Virtual Class "ADEg" [package "adegraphics"]
#> 
#> Slots:
#>                                                                        
#> Name:   trellis.par     adeg.par lattice.call       g.args        stats
#> Class:         list         list         list         list         list
#>                                 
#> Name:        s.misc         Call
#> Class:         list         call
#> 
#> Extends: "ADEgORtrellis", "ADEgORADEgSORtrellis"
#> 
#> Known Subclasses: 
#> Class "ADEg.C1", directly
#> Class "ADEg.S1", directly
#> Class "ADEg.S2", directly
#> Class "ADEg.T", directly
#> Class "ADEg.Tr", directly
#> Class "C1.barchart", by class "ADEg.C1", distance 2
#> Class "C1.curve", by class "ADEg.C1", distance 2
#> Class "C1.density", by class "ADEg.C1", distance 2
#> Class "C1.gauss", by class "ADEg.C1", distance 2
#> Class "C1.dotplot", by class "ADEg.C1", distance 2
#> Class "C1.hist", by class "ADEg.C1", distance 2
#> Class "C1.interval", by class "ADEg.C1", distance 2
#> Class "S1.boxplot", by class "ADEg.S1", distance 2
#> Class "S1.class", by class "ADEg.S1", distance 2
#> Class "S1.distri", by class "ADEg.S1", distance 2
#> Class "S1.label", by class "ADEg.S1", distance 2
#> Class "S1.match", by class "ADEg.S1", distance 2
#> Class "S2.arrow", by class "ADEg.S2", distance 2
#> Class "S2.class", by class "ADEg.S2", distance 2
#> Class "S2.corcircle", by class "ADEg.S2", distance 2
#> Class "S2.density", by class "ADEg.S2", distance 2
#> Class "S2.distri", by class "ADEg.S2", distance 2
#> Class "S2.image", by class "ADEg.S2", distance 2
#> Class "S2.label", by class "ADEg.S2", distance 2
#> Class "S2.logo", by class "ADEg.S2", distance 2
#> Class "S2.match", by class "ADEg.S2", distance 2
#> Class "S2.traject", by class "ADEg.S2", distance 2
#> Class "S2.value", by class "ADEg.S2", distance 2
#> Class "T.image", by class "ADEg.T", distance 2
#> Class "T.value", by class "ADEg.T", distance 2
#> Class "C1.curves", by class "ADEg.C1", distance 3
#> Class "T.cont", by class "ADEg.T", distance 3
#> Class "Tr.class", by class "ADEg.Tr", distance 2
#> Class "Tr.label", by class "ADEg.Tr", distance 2
#> Class "Tr.match", by class "ADEg.Tr", distance 2
#> Class "Tr.traject", by class "ADEg.Tr", distance 2