Class "ADEgS"

An object of ADEgS class is a complex graphic. This class allows the superposition, the insertion and/or the juxtaposition of several ADEg, trellis and/or ADEgS objects.

The ADEgS class have neither father class nor son classes.

Objects from the Class

ADEgS objects can be created by calls of the form new("ADEgS", ...).

The regular usage in this package is to use the ADEgS, add.ADEg, superpose, insert or + functions.

Slots

ADEglist

a list of several trellis, ADEg and/or ADEgS objects.

positions

a matrix with four columns and as many rows as the number of graphical objects in the ADEglist slot. For each simple graphic, i.e. in each row, the coordinates of the top-right and the bottom-left hand corners are in npc unit (normalized parent coordinates).

add

a square matrix with as many rows and columns as the number of graphical objects in the ADEglist slot. The value at the i-th row and j-th column is equal to 1 whether the j-th graphical object in the ADEglist slot is superpose to i-th graphical one. Otherwise, this value is equal to 0.

Call

an object of class call

Methods

[

signature(x = "ADEgS", i = "numeric", j = "missing", drop = "logical"): extracts the i-th sub-graphics in the x@ADEglist. i can be a vector. If i is a single number and if the extracted graphic in an ADEg object, the sub-selection is in the form of ADEg if drop is TRUE and in the form of ADEgS otherwise.

[

signature(x = "ADEgS", i = "numeric", j = "missing", drop = "missing"): the same than the previous method. drop is FALSE by default

[[

signature(x = "ADEgS", i = "numeric", j = "missing"): extracts one sub-graphic, the i-th one, in the x@ADEglist

[[

signature(x = "ADEgS", i = "character", j = "missing"): extracts one sub-graphic, named i in the x@ADEglist

[[<-

signature(x = "ADEgS", i = "numeric", j = "missing", value = "ADEg"): replaces one sub graphic, the i-th one, by an ADEg object in the x@ADEglist

[[<-

signature(x = "ADEgS", i = "numeric", j = "missing", value = "ADEgS"): replaces one sub graphic, the i-th one, by an ADEgS object in the x@ADEglist

$

signature(x = "ADEgS"): extracts one sub-graphic by its name in the x@ADEglist

getpositions

signature(object = "ADEgS"): returns the positions matrix of the object, i.e. object@positions

getgraphics

signature(object = "ADEgS"): returns the list of graphics of the object, i.e. object@ADEglist

getcall

signature(object = "ADEgS"): returns the call of the object, i.e. object@Call

names

signature(object = "ADEgS"): returns the graphics' names of the object, i.e. the names of object@ADEglist

names<-

signature(object = "ADEgS"): replaces the graphics' names of the object, i.e. the names of object@ADEglist

length

signature(x = "ADEgS"): returns the number of graphics into x, i.e. the length of x@ADEglist

plot

signature(x = "ADEgS"): same as print

print

signature(x = "ADEgS"): displays the graphical elements into one device using positions and superposition management (x@add matrix)

show

signature(object = "ADEgS"): same as print

superpose

signature(g1 = "ADEgS", g2 = "ADEg", which = "numeric", plot = "logical"): creates a new "ADEgS" object performing a superposition of g2 on the which-th ADEg object of g1. This object is printed if plot is TRUE.

superpose

signature(g1 = "ADEgS", g2 = "ADEg", which = "numeric", plot = "ANY"): creates a new "ADEgS" object performing a superposition of g2 on the which-th ADEg object of g1. This object is printed only if plot is TRUE.

superpose

signature(g1 = "ADEgS", g2 = "ADEg", which = "missing", plot = "ANY"): creates a new "ADEgS" object performing a superposition of g2 on the last ADEg object of g1. This object is printed only if plot is TRUE.

superpose

signature(g1 = "ADEgS", g2 = "ADEgS", which = "missing", plot = "ANY"): creates a new "ADEgS" object performing a superposition between two ADEgS having the same length and the same positions slot. It is used when g1 and g2 are both created with a partition of individual groups, variables or analysis' axis.

+

signature(e1 = "ADEg", e2 = "ADEgS"): creates a new "ADEgS" object performing a superposition of e1 on e2.

+

signature(e1 = "ADEgS", e2 = "ADEg"): creates a new "ADEgS" object performing a superposition of e2 on e1.

cbindADEg

signature(g1 = "ADEgORADEgS", g2 = "ADEgORADEgS"): creates a new "ADEgS" object combining g1 on g2 by columns.

rbindADEg

signature(g1 = "ADEgORADEgS", g2 = "ADEgORADEgS"): creates a new "ADEgS" object combining g1 on g2 by rows.

update

signature(object = "ADEgS"): modifies the graphical parameters of each sub-graphics listed in object@ADEglist and/or the object's names (with the key word names) and/or the object@positions slot (with the key word positions), after creation of the object. The current display is updated and a modified object is returned.

insert

signature(graphics = "ADEgS", oldgraphics = "missing", posi, ratio, inset, plot, which, dispatch): creates a new "ADEgS" object performing an insertion of graphics into the current device.

insert

signature(graphics = "ADEgS", oldgraphics = "ADEg", posi, ratio, inset, plot): creates a new "ADEgS" object performing an insertion of graphics into oldgraphics.

insert

signature(graphics = "ADEgORtrellis", oldgraphics = "ADEgS", posi, ratio, inset, plot, which): creates a new "ADEgS" object performing an insertion of graphics into oldgraphics.

insert

signature(graphics = "ADEgS", oldgraphics = "ADEgS", posi, ratio, inset, plot, which, dispatch): creates a new "ADEgS" object performing an insertion of graphics into oldgraphics.

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

See also

Examples

  showClass("ADEgS")
#> Class "ADEgS" [package "adegraphics"]
#> 
#> Slots:
#>                                               
#> Name:   ADEglist positions       add      Call
#> Class:      list    matrix    matrix      call
#> 
#> Extends: "ADEgORADEgSORtrellis"