Methods to display the outputs of an analysis performed with ade4

S3 methods to display the outputs of an analysis performed with ade4

Usage

# S3 method for foucart
kplot(object, xax = 1, yax = 2, which.tab = 1:length(object$blo), pos = -1, 
  storeData = TRUE, plot = TRUE, ...)
# S3 method for mcoa
kplot(object, xax = 1, yax = 2, which.tab = 1:nrow(object$cov2), 
  option = c("points", "axis", "columns"), pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for mfa
kplot(object, xax = 1, yax = 2, which.tab = 1:length(object$blo), traject = FALSE, 
  permute = FALSE, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for mbpcaiv
kplot(object, xax = 1, yax = 2, which.tab =
1:length(object$blo), pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for pta
kplot(object, xax = 1, yax = 2, which.tab = 1:nrow(object$RV), which.graph = 1:4, 
  pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for sepan
kplot(object, xax = 1, yax = 2, which.tab = 1:length(object$blo), permute = FALSE, 
  traject = FALSE, posieig = "bottomleft", pos = -1, storeData = TRUE, plot = TRUE, ...)
kplotsepan.coa(object, xax = 1, yax = 2, which.tab = 1:length(object$blo), 
  permute = FALSE, posieig = "bottomleft", pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for statis
kplot(object, xax = 1, yax = 2, which.tab = 1:length(object$tab.names), traject = FALSE, 
  arrow = TRUE, class = NULL, pos = -1, storeData = TRUE, plot = TRUE, ...)

# S3 method for acm
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for betcoi
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for betdpcoa
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for betwitdpcoa
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for betrlq
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for between
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for coinertia
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for discrimin
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for dpcoa
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for fca
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for foucart
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE,
plot = TRUE, ...)
# S3 method for krandboot
plot(x, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for krandxval
plot(x, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for mcoa
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for mfa
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE,
plot = TRUE, ...)
# S3 method for multiblock
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for multispati
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for niche
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for pcaiv
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for pta
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for procuste
plot(x, xax = 1, yax = 2, pos = -1, storeData =
TRUE, plot = TRUE, ...)
# S3 method for randboot
plot(x, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for randxval
plot(x, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for rlq
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for sepan
plot(x, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for statis
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for witcoi
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for witdpcoa
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for within
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for witrlq
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for bcaloocv
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)
# S3 method for discloocv
plot(x, xax = 1, yax = 2, pos = -1, storeData = TRUE, plot = TRUE, ...)

# S3 method for dudi
scatter(x, xax = 1, yax = 2, permute = FALSE, posieig = "topleft", prop = FALSE, 
  density.plot = ifelse(permute, ncol(x$tab) > 1000, nrow(x$tab) > 1000), plot = TRUE, 
  storeData = TRUE, pos = -1, ...)
# S3 method for coa
scatter(x, xax = 1, yax = 2, method = 1:3, posieig = "topleft", pos = -1, 
  storeData = TRUE, plot = TRUE, ...)
# S3 method for pco
scatter(x, xax = 1, yax = 2, posieig = "topleft", pos = -1, storeData = TRUE, 
  plot = TRUE, ...)
# S3 method for nipals
scatter(x, xax = 1, yax = 2, posieig = "topleft", pos = -1, storeData = TRUE, 
  plot = TRUE, ...)

# S3 method for acm
score(x, xax = 1, which.var = NULL, type = c("points", "boxplot"), pos = -1, 
  storeData = TRUE, plot = TRUE, ...)
# S3 method for mix
score(x, xax = 1, which.var = NULL, type = c("points", "boxplot"), pos = -1, 
  storeData = TRUE, plot = TRUE, ...)

# S3 method for pca
score(x, xax = 1, which.var = NULL, pos = -1, storeData = TRUE, plot = TRUE, ...)

# S3 method for dudi
screeplot(x, col.kept = "grey", col = "white", pos = -1, plot = TRUE, ...)

# S3 method for dudi
biplot(x, pos = -1, plot = TRUE, ...)

Arguments

object, x

objects used to select a method

xax

an integer (or a vector) indicating which column(s) of object or x is(are) plotted on the x-axis

yax

an integer (or a vector) indicating which column(s) of object or x is(are) plotted on the y-axis

which.tab

a numeric vector (used in kplot.*) containing the numbers of the tables used for the analysis

option

a string of characters (only used in kplot.mfa) indicating the drawing option: points plot of the projected scattergram onto the co-inertia axes, axis projections of inertia axes onto the co-inertia axes, columns projections of variables onto the synthetic variables planes.

which.graph

an integer between 1 and 4 (only used in kplot.pta) indicating the drawing option. For each table of which.tab, are drawn: 1 the projections of the principal axes, 2 the projections of the rows, 3 the projections of the columns, 4 the projections of the principal components onto the planes of the compromise.

permute

a logical value (used in kplot.sepan, kplotsepan.coa and scatter.dudi). If FALSE, the rows are plotted by points or density surface and the columns by arrows. If TRUE, it is the opposite.

traject

a logical value (used in kplot.sepan and kplot.statis) indicating whether the trajectories between rows should be drawn in a natural order

posieig

a character value or a two-length numeric vector (in normalized parent coordinates npc from 0 to 1) or none value indicating the position of the eigenvalues bar plot (used in kplot.sepan, kplotsepan.coa and scatter.*).

arrow

a logical value (only used in kplot.statis) indicating whether the column factorial diagrams should be plotted

class

if not NULL, a factor of length equal to the number of the total columns of the K-tables (only used in kplot.statis)

prop

a logical value (only used in scatter.dudi) indicating if the size of the arrows' labels is proportional to the analysis score.

density.plot

a logical value (only used in scatter.dudi)indicating if the points are displayed as density surface (using s.density).

method

an integer between 1 and 3 (only used in scatter.coa) indicating the drawing option. Are drawn: 1 rows and columns with the coordinates of lambda variance, 2 rows variance 1 and columns by averaging, 3 columns variance 1 and rows by averaging.

which.var

the numbers of the kept columns for the analysis, otherwise all columns (used in score.*)

type

a string of characters (only used in score.acm and score.mix) indicating if points (points) or boxplot (boxplot) are used to represent levels of factors

col.kept

one color value to color the kept axes in the barchart (used in screeplot.dudi)

col

one color value to color the axes in the barchart (used in screeplot.dudi)

plot

a logical indicating if the graphics is displayed

storeData

a logical indicating if the data should be stored in the returned object. If FALSE, only the names of the data arguments are stored

pos

an integer indicating the position of the environment where the data are stored, relative to the environment where the function is called. Useful only if storeData is FALSE

...

additional graphical parameters (see adegpar and trellis.par.get)

Value

Returns an ADEg or an ADEgS object. The result is displayed if plot is TRUE.

Author

Alice Julien-Laferriere, Aurelie Siberchicot aurelie.siberchicot@univ-lyon1.fr and Stephane Dray

References

See ade4 website: <URL: http://pbil.univ-lyon1.fr/ADE-4/>

Examples

cat("To run the example on 'topic'\n")
#> To run the example on 'topic'
cat("Type in your R console: example(topic, package = 'ade4') \n")
#> Type in your R console: example(topic, package = 'ade4')